Don't kill worker if run @pid is empty.

+++ app/models/run.coffee
@@ -140,7 +140,8 @@ schema.methods.stoppable = ->
   @status in ['idle', 'busy']
 
 schema.methods.stop = ->
-  process.kill @pid, 'SIGINT'
+  if @pid
+    process.kill @pid, 'SIGINT'
 
 schema.methods.setProgress = (current, max, callback) ->
   if current == 'max'
